% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_task_classif_st.R
\name{as_task_classif.TaskClassifST}
\alias{as_task_classif.TaskClassifST}
\title{Convert to a Classification Task}
\usage{
\method{as_task_classif}{TaskClassifST}(x, ...)
}
\arguments{
\item{x}{(any)\cr
Object to convert.}

\item{...}{(any)\cr
Additional arguments.}
}
\description{
Convert to a Classification Task
}
