% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_task_classif_st.R
\name{as_task_classif_st}
\alias{as_task_classif_st}
\alias{as_task_classif_st.TaskClassifST}
\alias{as_task_classif_st.data.frame}
\alias{as_task_classif_st.DataBackend}
\alias{as_task_classif_st.sf}
\title{Convert to a SpatioTemporal Classification Task}
\usage{
as_task_classif_st(x, ...)

\method{as_task_classif_st}{TaskClassifST}(x, clone = FALSE, ...)

\method{as_task_classif_st}{data.frame}(
  x,
  target = NULL,
  id = deparse(substitute(x)),
  positive = NULL,
  coordinate_names = c("x", "y"),
  ...
)

\method{as_task_classif_st}{DataBackend}(
  x,
  target = NULL,
  id = deparse(substitute(x)),
  positive = NULL,
  coordinate_names = c("x", "y"),
  ...
)

\method{as_task_classif_st}{sf}(
  x,
  target = NULL,
  id = deparse(substitute(x)),
  positive = NULL,
  ...
)
}
\arguments{
\item{x}{(\code{any})\cr
Object to convert.}

\item{...}{(\code{any})\cr
Additional arguments.}

\item{clone}{(\code{logical(1)})\cr
If \code{TRUE}, ensures that the returned object is not the same as the input \code{x}.}

\item{target}{(\code{character(1)})\cr
Name of the target column.}

\item{id}{(\code{character(1)})\cr
Id for the new task.
Defaults to the (deparsed and substituted) name of \code{x}.}

\item{positive}{(\code{character(1)})\cr
Level of the positive class. See \link[mlr3]{TaskClassif}.}

\item{coordinate_names}{(\code{character()})\cr
Name of the coordinate columns. Defaults to \code{c("x", "y")}.}
}
\value{
\link{TaskClassifST}.
}
\description{
Convert object to a \link{TaskClassifST}.
This is a S3 generic, specialized for at least the following objects:
\enumerate{
\item \link{TaskClassifST}: ensure the identity
\item \code{\link[=data.frame]{data.frame()}} and \link{DataBackend}: provides an alternative to the constructor of \link{TaskClassifST}.
\item \link[sf:sf]{sf::sf}.
}
}
\examples{
data("ecuador", package = "mlr3spatiotempcv")
as_task_classif_st(ecuador, target = "slides", positive = "TRUE")

ecuador_sf = sf::st_as_sf(ecuador, coords = c("x", "y"), crs = 4326)
as_task_classif_st(ecuador_sf, target = "slides", positive = "TRUE")
}
