% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_task_regr_st.R
\name{as_task_regr}
\alias{as_task_regr}
\alias{as_task_regr.TaskRegrST}
\title{Convert to a Regression Task}
\usage{
as_task_regr(x, ...)

\method{as_task_regr}{TaskRegrST}(x)
}
\arguments{
\item{x}{(\code{any})\cr
Object to convert.}

\item{...}{(\code{any})\cr
Additional arguments.}
}
\value{
\link[mlr3]{TaskRegr}.
}
\description{
Convert object to a \link{TaskRegr}.
This is a S3 generic for \link{TaskRegrST}.
}
\seealso{
\code{\link[mlr3:as_task_regr]{mlr3::as_task_regr()}}
}
