% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_DataBackend.R
\name{assert_spatial_task}
\alias{assert_spatial_task}
\title{Check spatial task}
\usage{
assert_spatial_task(task)
}
\arguments{
\item{task}{(\link[mlr3]{Task}).}
}
\description{
Assertion helper for spatial mlr3 tasks.
}
\keyword{internal}
