% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\name{autoplot}
\alias{autoplot}
\title{Re-export of \code{autoplot}

See \code{\link[ggplot2:autoplot]{ggplot2::autoplot()}}.}
\description{
Re-export of \code{autoplot}

See \code{\link[ggplot2:autoplot]{ggplot2::autoplot()}}.
}
\keyword{internal}
