% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VizFold.R
\name{autoplot.ResamplingSpCVBuffer}
\alias{autoplot.ResamplingSpCVBuffer}
\title{Plot for Spatial Resampling}
\usage{
\method{autoplot}{ResamplingSpCVBuffer}(object, task, fold_id, ...)
}
\arguments{
\item{object}{\link[mlr3spatiotemporal:ResamplingSpCVBuffer]{mlr3spatiotemporal::ResamplingSpCVBuffer}.}

\item{task}{\link[mlr3spatiotemporal:TaskClassifST]{mlr3spatiotemporal::TaskClassifST}}

\item{fold_id}{\code{\link[=numeric]{numeric()}}}

\item{...}{Currently not used}
}
\value{
\code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object.
}
\description{
Generates plots for \link[mlr3spatiotemporal:ResamplingSpCVBuffer]{mlr3spatiotemporal::ResamplingSpCVBuffer}.
}
\examples{
library(mlr3)
library(ggplot2)
task <- mlr3::mlr_tasks$get("ecuador")
resampling <- ResamplingSpCVBuffer$new()
resampling$param_set$values <- list(range = 1000)
resampling$instantiate(task)
autoplot(resampling, task, 1)
}
