% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VizFold.R
\name{autoplot.ResamplingSpCVCoords}
\alias{autoplot.ResamplingSpCVCoords}
\title{Plot for Spatial Resampling}
\usage{
\method{autoplot}{ResamplingSpCVCoords}(object, task, fold_id = NULL,
  ...)
}
\arguments{
\item{object}{\link[mlr3spatiotemporal:ResamplingSpCVCoords]{mlr3spatiotemporal::ResamplingSpCVCoords}}

\item{task}{\link[mlr3spatiotemporal:TaskClassifST]{mlr3spatiotemporal::TaskClassifST}}

\item{fold_id}{\code{\link[=numeric]{numeric()}}}

\item{...}{Currently not used}
}
\value{
\code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object.
}
\description{
Generates plots for \link[mlr3spatiotemporal:ResamplingSpCVCoords]{mlr3spatiotemporal::ResamplingSpCVCoords}.
}
\examples{
library(mlr3)
library(ggplot2)
task = mlr3::mlr_tasks$get("ecuador")
resampling = ResamplingSpCVCoords$new()
resampling$param_set$values = list(folds = 4)
resampling$instantiate(task)
autoplot(resampling, task)
autoplot(resampling, task, 1)
autoplot(resampling, task, c(1, 2, 3, 4))
}
