% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.R
\name{autoplot_spatial}
\alias{autoplot_spatial}
\title{Plot for Spatial Resampling}
\usage{
autoplot_spatial(resampling, task, fold_id = NULL, grid = TRUE)
}
\arguments{
\item{resampling}{\link{ResamplingSpCVCoords}}

\item{task}{\link{TaskClassifST}}

\item{fold_id}{\code{\link[=numeric]{numeric()}}}

\item{...}{Currently not used}
}
\value{
\code{\link[=ggplot]{ggplot()}} resampling.
}
\description{
Workhorse for autoplot.ResamplingSpCV*.
}
\keyword{internal}
