% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.R
\name{autoplot.ResamplingSpCVBlock}
\alias{autoplot.ResamplingSpCVBlock}
\alias{autoplot.ResamplingSpCVEnv}
\alias{autoplot.ResamplingSpCVBuffer}
\alias{autoplot.ResamplingSpCVCoords}
\title{Visualization of spatial resampling.}
\usage{
\method{autoplot}{ResamplingSpCVBlock}(object, task, fold_id = NULL,
  grid = TRUE, ...)

\method{autoplot}{ResamplingSpCVEnv}(object, task, fold_id = NULL,
  grid = TRUE, ...)

\method{autoplot}{ResamplingSpCVBuffer}(object, task, fold_id = NULL,
  grid = TRUE, ...)

\method{autoplot}{ResamplingSpCVCoords}(object, task, fold_id = NULL,
  grid = TRUE, ...)
}
\arguments{
\item{object}{\link{ResamplingSpCVBlock}\cr
An instantianated spatial resampling.}

\item{task}{\link{TaskClassifST}\cr
A spatial task.}

\item{fold_id}{\code{\link[=numeric]{numeric()}}\cr
For which fold ids should training and test sets be visualized?}

\item{grid}{(\code{logical(1)})\cr
Should a gridded plot using \code{cowplot::plot_grid()} be created? If ´FALSE\code{only a list with all ggplot2 resamplings is returned. Default is}TRUE\code{. Only applies if a numeric vector is passed to argument }fold_id`.}

\item{...}{Currently not used}
}
\value{
\code{\link[=ggplot]{ggplot()}} resampling.

\code{\link[=ggplot]{ggplot()}} resampling.

\code{\link[=ggplot]{ggplot()}} resampling.

\code{\link[=ggplot]{ggplot()}} resampling.
}
\description{
Generates plots for \link{ResamplingSpCVBlock}, \link{ResamplingSpCVEnv},
\link{ResamplingSpCVBuffer} and \link{ResamplingSpCVCoords}.

Generates plots for \link[mlr3spatiotemporal:ResamplingSpCVEnv]{mlr3spatiotemporal::ResamplingSpCVEnv}.

Generates plots for \link[mlr3spatiotemporal:ResamplingSpCVCoords]{mlr3spatiotemporal::ResamplingSpCVCoords}.
}
\details{
By default a plot is returned; if \code{fold_id} is set, a gridded plot
is created. If \code{grid = FALSE}, only a list of ggplot2 resamplings is returned.
This gives the option to align the single plots manually.
}
\examples{
#####
# SpCVBlock
#####
\dontrun{
library(mlr3)
task <- tsk("ecuador")
resampling = rsmp("spcv-block")
resampling$param_set$values = list(folds = 4)
resampling$instantiate(task)

## Visualize all partitions
autoplot(resampling, task)

## Visualize the train/test split of a single fold
autoplot(resampling, task, fold_id = 1)

## Visualize train/test splits of multiple folds
autoplot(resampling, task, fold_id = c(1, 2, 3, 4))

# return only a list of ggplot2 resamplings
plot_list = autoplot(resampling, task,
  fold_id = c(1, 2, 3, 4), grid = FALSE)
}
#####
# SpCVEnv
#####
\dontrun{
resampling = rsmp("spcv-env")
resampling$param_set$values = list(folds = 4, features = c("dem"))
resampling$instantiate(task)
autoplot(resampling, task)
autoplot(resampling, task, 1)
autoplot(resampling, task, c(1, 2, 3, 4))
}
#####
# SpCVBuffer
#####
\dontrun{
resampling = rsmp("spcv-buffer")
resampling$param_set$values = list(range = 1000)
resampling$instantiate(task)
autoplot(resampling, task, 1)
autoplot(resampling, task, c(1, 2, 3, 4))
}
#####
# SpCVCoords
#####
\dontrun{
resampling = rsmp("spcv-coords")
resampling$instantiate(task)
autoplot(resampling, task)
autoplot(resampling, task, 1)
autoplot(resampling, task, c(1, 2, 3, 4))
}
}
