% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_DataBackend.R
\name{get_coordinate_names}
\alias{get_coordinate_names}
\title{Get coordinates names from Task or Backend}
\usage{
get_coordinate_names(task)
}
\arguments{
\item{task}{(\link[mlr3]{Task}).}
}
\description{
Returns the coordinates names of the given Task or Backend.
}
\keyword{internal}
