% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_DataBackend.R
\name{get_coordinates}
\alias{get_coordinates}
\title{Get coordinates from Task or Backend}
\usage{
get_coordinates(task)
}
\arguments{
\item{task}{(\link[mlr3]{Task}).}
}
\description{
Returns the coordinates of the given Task or Backend.
}
\keyword{internal}
