% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_DataBackend.R
\name{get_crs}
\alias{get_crs}
\title{Get CRS from Task or Backend}
\usage{
get_crs(task, pretty = TRUE)
}
\arguments{
\item{task}{(\link[mlr3]{Task}).}

\item{pretty}{\verb{[logical]}\cr
If \code{TRUE}, returns the pretty coordinate information instead of raw CRS
string.}
}
\description{
Returns the CRS of the given Task or Backend.
}
\keyword{internal}
