% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr3spatiotempcv-package.R, R/zzz.R
\docType{package}
\name{mlr3spatiotempcv-package}
\alias{mlr3spatiotempcv}
\alias{mlr3spatiotempcv-package}
\title{mlr3spatiotempcv: Spatiotemporal Resampling Methods for 'mlr3'}
\description{
Implements spatio-temporal methods to enhance the
    experience of mlr3.

Implements spatio-temporal methods to enhance the
    experience of mlr3.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://mlr3spatiotempcv.mlr-org.com/}
  \item \url{https://github.com/mlr-org/mlr3spatiotempcv}
  \item Report bugs at \url{https://github.com/mlr-org/mlr3spatiotempcv/issues}
}


Useful links:
\itemize{
  \item \url{https://mlr3spatiotempcv.mlr-org.com/}
  \item \url{https://github.com/mlr-org/mlr3spatiotempcv}
  \item Report bugs at \url{https://github.com/mlr-org/mlr3spatiotempcv/issues}
}

}
\author{
\strong{Maintainer}: Patrick Schratz \email{patrick.schratz@gmail.com} (\href{https://orcid.org/0000-0003-0748-6624}{ORCID})

Authors:
\itemize{
  \item Marc Becker \email{marcbecker@posteo.de}
}

Other contributors:
\itemize{
  \item Jannes Muenchow \email{jannes.muenchow@uni-jena.de} [contributor]
}


\strong{Maintainer}: Patrick Schratz \email{patrick.schratz@gmail.com} (\href{https://orcid.org/0000-0003-0748-6624}{ORCID})

Authors:
\itemize{
  \item Marc Becker \email{marcbecker@posteo.de}
}

Other contributors:
\itemize{
  \item Jannes Muenchow \email{jannes.muenchow@uni-jena.de} [contributor]
}

}
\keyword{internal}
