% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mlr3spatiotemporal-package.R, R/zzz.R
\docType{package}
\name{mlr3spatiotemporal-package}
\alias{mlr3spatiotemporal}
\alias{mlr3spatiotemporal-package}
\title{mlr3spatiotemporal: Spatial support for mlr3}
\description{
Implements spatio-temporal methods to enhance the
    experience of mlr3.

Implements spatio-temporal methods to enhance the
    experience of mlr3.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://mlr3spatiotemporal.mlr-org.com/}
  \item \url{https://github.com/mlr-org/mlr3spatiotemporal}
  \item Report bugs at \url{https://github.com/mlr-org/mlr3spatiotemporal/issues}
}


Useful links:
\itemize{
  \item \url{https://mlr3spatiotemporal.mlr-org.com/}
  \item \url{https://github.com/mlr-org/mlr3spatiotemporal}
  \item Report bugs at \url{https://github.com/mlr-org/mlr3spatiotemporal/issues}
}

}
\author{
\strong{Maintainer}: Patrick Schratz \email{patrick.schratz@gmail.com} (0000-0003-0748-6624)


\strong{Maintainer}: Patrick Schratz \email{patrick.schratz@gmail.com} (0000-0003-0748-6624)

}
\keyword{internal}
