% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Task_classif_diplodia.R
\name{mlr_tasks_diplodia}
\alias{mlr_tasks_diplodia}
\title{Diplodia Classification Task}
\format{\link[R6:R6Class]{R6::R6Class} inheriting from \link{TaskClassif}.}
\description{
Data set created by Patrick Schratz, University of Jena
(Germany) and Eugenia Iturritxa, NEIKER, Vitoria-Gasteiz (Spain). This
dataset should be cited as Schratz et al. (2019) (see reference below). The
publication also contains additional information on data collection. The
data set provided here shows infections of trees by the pathogen \emph{Diplodia
Sapinea} in the Basque Country in Spain. Predictors are environmental
variables like temperature, precipitation, soil and more.
}
\section{Usage}{
\preformatted{mlr_tasks$get("diplodia")
}
}

\references{
Schratz, P., Muenchow, J., Iturritxa, E., Richter, J., &
Brenning, A. (2019). Hyperparameter tuning and performance assessment of
statistical and machine-learning algorithms using spatial data. Ecological
Modelling, 406, 109–120. https://doi.org/10/gf34bd
}
