% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Task_classif_ecuador.R
\docType{data}
\name{mlr_tasks_ecuador}
\alias{mlr_tasks_ecuador}
\alias{ecuador}
\title{Ecuador Classification Task}
\format{
\link[R6:R6Class]{R6::R6Class} inheriting from \link{TaskClassif}.
}
\usage{
data(ecuador)
}
\description{
Data set created by Jannes Muenchow, University of Erlangen-Nuernberg,
Germany. This dataset should be cited as Muenchow et al. (2012) (see reference
below). The publication also contains additional information on data
collection and the geomorphology of the area. The data set provided here is (a
subset of) the one from the 'natural' part of the RBSF area and corresponds
to landslide distribution in the year 2000.
}
\section{Usage}{
\preformatted{mlr_tasks$get("ecuador")
tsk("ecuador")
}
}

\references{
Muenchow, J., Brenning, A., Richter, M., 2012. Geomorphic process
rates of landslides along a humidity gradient in the tropical Andes.
Geomorphology, 139-140: 271-284.
}
\seealso{
\link[mlr3misc:Dictionary]{Dictionary} of \link[=Task]{Tasks}: \link{mlr_tasks}

\code{as.data.table(mlr_tasks)} for a complete table of all (also dynamically created) \link[=Task]{Tasks}.

Other Task: 
\code{\link{TaskClassifST}},
\code{\link{TaskRegrST}},
\code{\link{mlr_tasks_cookfarm}},
\code{\link{mlr_tasks_diplodia}}
}
\concept{Task}
\keyword{datasets}
