% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VizFold.R
\name{plot.ResamplingSpCVBlock}
\alias{plot.ResamplingSpCVBlock}
\title{Plot for Spatial Resampling}
\usage{
\method{plot}{ResamplingSpCVBlock}(object, task)
}
\arguments{
\item{object}{(\link[mlr3spatiotemporal:ResamplingSpCVBlock]{mlr3spatiotemporal::ResamplingSpCVBlock}).}
}
\value{
\code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object.
}
\description{
Generates plots for \link[mlr3spatiotemporal:ResamplingSpCVBlock]{mlr3spatiotemporal::ResamplingSpCVBlock}.
}
\examples{
library(mlr3)
task <- mlr3::mlr_tasks$get("ecuador")
resampling <- ResamplingSpCVBlock$new()
resampling$param_set$values <- list(folds = 4)
resampling$instantiate(task)
plot(resampling, task)
}
