% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VizFold.R
\name{plot.ResamplingSpCVBuffer}
\alias{plot.ResamplingSpCVBuffer}
\title{Plot for Spatial Resampling}
\usage{
\method{plot}{ResamplingSpCVBuffer}(object, task, fold)
}
\arguments{
\item{object}{(\link[mlr3spatiotemporal:ResamplingSpCVBuffer]{mlr3spatiotemporal::ResamplingSpCVBuffer}).}
}
\value{
\code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object.
}
\description{
Generates plots for \link[mlr3spatiotemporal:ResamplingSpCVBuffer]{mlr3spatiotemporal::ResamplingSpCVBuffer}.
}
\examples{
library(mlr3)
task <- mlr3::mlr_tasks$get("ecuador")
resampling <- ResamplingSpCVBlock$new()
resampling$param_set$values <- list(range = 1000)
resampling$instantiate(task)
plot(resampling, task)
}
