% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/VizFold.R
\name{plot.ResamplingSpCVKmeans}
\alias{plot.ResamplingSpCVKmeans}
\title{Plot for Spatial Resampling}
\usage{
\method{plot}{ResamplingSpCVKmeans}(object, task)
}
\arguments{
\item{object}{(\link[mlr3spatiotemporal:ResamplingSpCVKmeans]{mlr3spatiotemporal::ResamplingSpCVKmeans}).}
}
\value{
\code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object.
}
\description{
Generates plots for \link[mlr3spatiotemporal:ResamplingSpCVKmeans]{mlr3spatiotemporal::ResamplingSpCVKmeans}.
}
\examples{
library(mlr3)
task <- mlr3::mlr_tasks$get("ecuador")
resampling <- ResamplingSpCVKmeans$new()
resampling$param_set$values <- list(folds = 4)
resampling$instantiate(task)
plot(resampling, task)
}
