% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResamplingSpCVBuffer.R
\docType{data}
\name{ResamplingSpCVBuffer}
\alias{ResamplingSpCVBuffer}
\title{Spatial Buffer Cross Validation Resampling}
\format{\link[R6:R6Class]{R6::R6Class} inheriting from \link{Resampling}.}
\usage{
ResamplingSpCVBuffer
}
\description{
Spatial Buffer Cross validation implemented by the \code{blockCV} package.
}
\section{Fields}{

See \link{Resampling}.
}

\section{Methods}{

See \link{Resampling}.
}

\examples{
\dontrun{
library(mlr3)
task = tsk("ecuador")

# Instantiate Resampling
rcv = rsmp("spcv-buffer")
rcv$param_set$values = list(range = 1000)
rcv$instantiate(task)

# Individual sets:
rcv$train_set(1)
rcv$test_set(1)
intersect(rcv$train_set(1), rcv$test_set(1))

# Internal storage:
rcv$instance
}
}
\references{
Valavi R, Elith J, Lahoz-Monfort JJ, Guillera-Arroita G. blockCV: An r package for generating spatially or environmentally separated folds for k-fold cross-validation of species distribution models. Methods Ecol Evol. 2019; 10:225–232. https://doi.org/10.1111/2041-210X.13107
}
\keyword{datasets}
