% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResamplingSpCVCoords.R
\name{ResamplingSpCVCoords}
\alias{ResamplingSpCVCoords}
\title{Spatial Cross Validation Resampling}
\format{\link[R6:R6Class]{R6::R6Class} inheriting from \link{Resampling}.}
\description{
Spatial Cross validation following the "k-means" approach after
Brenning 2012.
}
\section{Fields}{
 See \link{Resampling}.
}

\section{Methods}{
 See \link{Resampling}.
}

\examples{
library(mlr3)
task = tsk("ecuador")

# Instantiate Resampling
rcv = rsmp("spcv-coords")
rcv$param_set$values = list(folds = 3)
rcv$instantiate(task)

# Individual sets:
rcv$train_set(1)
rcv$test_set(1)
intersect(rcv$train_set(1), rcv$test_set(1))

# Internal storage:
rcv$instance # table
}
\references{
Brenning, A. (2012). Spatial cross-validation and bootstrap for
the assessment of prediction rules in remote sensing: The R package
sperrorest. 2012 IEEE International Geoscience and Remote Sensing Symposium.
https://doi.org/10.1109/igarss.2012.6352393
}
