% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper_autoplot.R
\name{strat_sample_folds}
\alias{strat_sample_folds}
\title{Stratified random sampling}
\usage{
strat_sample_folds(data, col, n)
}
\arguments{
\item{data}{(\code{data.table})\cr
Data}

\item{col}{(\code{character(1)})\cr
Column to stratify on.}

\item{n}{(\code{integer})\cr
Sample size per group}
}
\description{
Stratified random sampling
}
\keyword{internal}
