% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.R, R/autoplot_SpCVBuffer.R,
%   R/autoplot_SptCVCstf.R
\name{autoplot_spatial_resampling}
\alias{autoplot_spatial_resampling}
\alias{autoplot.ResamplingSpCVBlock}
\alias{plot.ResamplingSpCVBlock}
\alias{autoplot.ResamplingSpCVEnv}
\alias{plot.ResamplingSpCVEnv}
\alias{autoplot.ResamplingSpCVCoords}
\alias{plot.ResamplingSpCVCoords}
\alias{autoplot.ResamplingSptCVCluto}
\alias{plot.ResamplingSptCVCluto}
\alias{autoplot.ResamplingCV}
\alias{plot.ResamplingCV}
\alias{autoplot.ResamplingRepeatedCV}
\alias{plot.ResamplingRepeatedCV}
\alias{autoplot.ResamplingRepeatedSpCVCoords}
\alias{plot.ResamplingRepeatedSpCVCoords}
\alias{autoplot.ResamplingRepeatedSpCVEnv}
\alias{plot.ResamplingRepeatedSpCVEnv}
\alias{autoplot.ResamplingRepeatedSpCVBlock}
\alias{plot.ResamplingRepeatedSpCVBlock}
\alias{autoplot.ResamplingRepeatedSptCVCluto}
\alias{plot.ResamplingRepeatedSptCVCluto}
\alias{autoplot.ResamplingSpCVBuffer}
\alias{plot.ResamplingSpCVBuffer}
\alias{autoplot.ResamplingSptCVCstf}
\alias{plot.ResamplingSptCVCstf}
\title{Visualization of spatial resampling.}
\usage{
\method{autoplot}{ResamplingSpCVBlock}(
  object,
  task,
  fold_id = NULL,
  grid = TRUE,
  train_color = "#0072B5",
  test_color = "#E18727",
  crs = NULL,
  ...
)

\method{plot}{ResamplingSpCVBlock}(x, ...)

\method{autoplot}{ResamplingSpCVEnv}(
  object,
  task,
  fold_id = NULL,
  grid = TRUE,
  train_color = "#0072B5",
  test_color = "#E18727",
  ...
)

\method{plot}{ResamplingSpCVEnv}(x, ...)

\method{autoplot}{ResamplingSpCVCoords}(
  object,
  task,
  fold_id = NULL,
  grid = TRUE,
  train_color = "#0072B5",
  test_color = "#E18727",
  crs = NULL,
  ...
)

\method{plot}{ResamplingSpCVCoords}(x, ...)

\method{autoplot}{ResamplingSptCVCluto}(
  object,
  task,
  fold_id = NULL,
  grid = TRUE,
  train_color = "#0072B5",
  test_color = "#E18727",
  tickformat_date = "\%Y-\%m",
  crs = NULL,
  nticks_x = 3,
  nticks_y = 3,
  point_size = 3,
  axis_label_fontsize = 11,
  ...
)

\method{plot}{ResamplingSptCVCluto}(x, ...)

\method{autoplot}{ResamplingCV}(
  object,
  task,
  fold_id = NULL,
  grid = TRUE,
  train_color = "#0072B5",
  test_color = "#E18727",
  ...
)

\method{plot}{ResamplingCV}(x, ...)

\method{autoplot}{ResamplingRepeatedCV}(
  object,
  task,
  fold_id = NULL,
  repeats_id = 1,
  grid = TRUE,
  train_color = "#0072B5",
  test_color = "#E18727",
  crs = NULL,
  ...
)

\method{plot}{ResamplingRepeatedCV}(x, ...)

\method{autoplot}{ResamplingRepeatedSpCVCoords}(
  object,
  task,
  fold_id = NULL,
  repeats_id = 1,
  grid = TRUE,
  train_color = "#0072B5",
  test_color = "#E18727",
  crs = NULL,
  ...
)

\method{plot}{ResamplingRepeatedSpCVCoords}(x, ...)

\method{autoplot}{ResamplingRepeatedSpCVEnv}(
  object,
  task,
  fold_id = NULL,
  repeats_id = 1,
  grid = TRUE,
  train_color = "#0072B5",
  test_color = "#E18727",
  crs = NULL,
  ...
)

\method{plot}{ResamplingRepeatedSpCVEnv}(x, ...)

\method{autoplot}{ResamplingRepeatedSpCVBlock}(
  object,
  task,
  fold_id = NULL,
  repeats_id = 1,
  grid = TRUE,
  train_color = "#0072B5",
  test_color = "#E18727",
  crs = NULL,
  ...
)

\method{plot}{ResamplingRepeatedSpCVBlock}(x, ...)

\method{autoplot}{ResamplingRepeatedSptCVCluto}(
  object,
  task,
  fold_id = NULL,
  repeats_id = 1,
  grid = TRUE,
  train_color = "#0072B5",
  test_color = "#E18727",
  tickformat_date = "\%Y-\%m",
  crs = NULL,
  nticks_x = 3,
  nticks_y = 3,
  point_size = 3,
  axis_label_fontsize = 11,
  ...
)

\method{plot}{ResamplingRepeatedSptCVCluto}(x, ...)

\method{autoplot}{ResamplingSpCVBuffer}(
  object,
  task,
  fold_id = NULL,
  grid = TRUE,
  train_color = "#0072B5",
  test_color = "#E18727",
  ...
)

\method{plot}{ResamplingSpCVBuffer}(x, ...)

\method{autoplot}{ResamplingSptCVCstf}(
  object,
  task,
  fold_id = NULL,
  repeats_id = NULL,
  grid = TRUE,
  train_color = "#0072B5",
  test_color = "#E18727",
  tickformat_date = "\%Y-\%m",
  crs = NULL,
  nticks_x = 3,
  nticks_y = 3,
  point_size = 3,
  axis_label_fontsize = 11,
  ...
)

\method{plot}{ResamplingSptCVCstf}(x, ...)
}
\arguments{
\item{object}{\verb{[Resampling]}\cr
mlr3 spatial resampling object. One of class \link{ResamplingSpCVBuffer},
\link{ResamplingSpCVBlock}, \link{ResamplingSpCVCoords}, \link{ResamplingSpCVEnv}.}

\item{task}{\verb{[TaskClassifST]/[TaskRegrST]}\cr
mlr3 task object.}

\item{fold_id}{\verb{[numeric]}\cr
Fold IDs to plot.}

\item{train_color}{\verb{[character(1)]}\cr
The color to use for the training set observations.}

\item{test_color}{\verb{[character(1)]}\cr
The color to use for the test set observations.}

\item{crs}{\verb{[character]}\cr
EPSG code of the CRS for x and y axes.
If not set, EPSG 4326 (WGS84) is used.}

\item{...}{Not used.}

\item{tickformat_date}{\verb{[character]}\cr
Date format for z-axis.}

\item{nticks_x}{\verb{[integer]}\cr
Number of x axis breaks.}

\item{nticks_y}{\verb{[integer]}\cr
Number of y axis breaks.}

\item{point_size}{\verb{[numeric]}\cr
Point size of markers.}

\item{axis_label_fontsize}{\verb{[integer]}\cr
Font size of axis labels.}

\item{repeats_id}{\verb{[numeric]}\cr
Repetition ID to plot.}

\item{plot_as_grid}{\verb{[logical(1)]}\cr
Should a gridded plot using \code{cowplot::plot_grid()} be created? If \code{FALSE}
only a list with all \CRANpkg{ggplot2} resamplings is returned. Default is
\code{TRUE}. Only applies if a numeric vector is passed to argument \code{fold_id}.}
}
\value{
\code{\link[=ggplot]{ggplot()}} or list of ggplot2 objects.
}
\description{
Generates plots for \link{ResamplingSpCVBlock}, \link{ResamplingSpCVEnv},
\link{ResamplingSpCVBuffer} and \link{ResamplingSpCVCoords}.
}
\details{
By default a plot is returned; if \code{fold_id} is set, a gridded plot is
created. If \code{plot_as_grid = FALSE}, a list of plot objects is returned.
This can be used to align the plots individually.

When no single fold is selected, the \code{\link[ggsci:scale_ucscgb]{ggsci::scale_color_ucscgb()}} palette
is used to display all partitions.
If you want to change the colors, call \verb{<plot> + <color-palette>()}.
}
\examples{
##########
# SpCVBlock
##########
\dontrun{
library(mlr3)
task = tsk("ecuador")
resampling = rsmp("spcv-block", range = 1000)
resampling$instantiate(task)

## Visualize all partitions
autoplot(resampling, task)

## Visualize the train/test split of a single fold
autoplot(resampling, task, fold_id = 1)

## Visualize train/test splits of multiple folds
autoplot(resampling, task, fold_id = c(1, 2, 3, 4))

# list of ggplot2 resamplings
plot_list = autoplot(resampling, task,
  fold_id = c(1, 2, 3, 4), grid = FALSE)
}
##########
# SpCVEnv
##########
\dontrun{
library(mlr3)
resampling = rsmp("spcv-env", folds = 4, features = "dem")
resampling$instantiate(task)

autoplot(resampling, task)
autoplot(resampling, task, 1)
autoplot(resampling, task, c(1, 2, 3, 4))
}
##########
# SpCVCoords
##########
\dontrun{
library(mlr3)
resampling = rsmp("spcv-coords")
resampling$instantiate(task)

autoplot(resampling, task)
autoplot(resampling, task, 1)
autoplot(resampling, task, c(1, 2, 3, 4))
}
##########
# SptCVCluto
##########
\dontrun{
library(mlr3)
library(mlr3spatiotempcv)
task_st = tsk("cookfarm")
resampling = rsmp("sptcv-cluto", folds = 5)
resampling$instantiate(task_st, "Date")

# plot
# resampling = readRDS("inst/spt-plotting-data.rda")
autoplot(resampling, task_st)
autoplot(resampling, task_st, 1)
autoplot(resampling, task_st, c(1, 2, 3, 4))
}
##########
# Non-Spatial CV
##########
\dontrun{
library(mlr3)
resampling = rsmp("cv")
resampling$instantiate(task)

autoplot(resampling, task)
autoplot(resampling, task, 1)
autoplot(resampling, task, c(1, 2, 3, 4))
}
##########
# Repeated Non-Spatial CV
##########
\dontrun{
library(mlr3)
resampling = rsmp("repeated_cv", folds = 5, repeats = 2)
resampling$instantiate(task)

autoplot(resampling, task)
autoplot(resampling, task, 1)
autoplot(resampling, task, fold_id = 2, repeats_id = 2)
autoplot(resampling, task, c(1, 2, 3, 4))
}
##########
# RepeatedSpCVCoords
##########
\dontrun{
library(mlr3)
task = tsk("diplodia")
resampling = rsmp("repeated-spcv-coords", folds = 10, repeats = 2)
resampling$instantiate(task)

autoplot(resampling, task)
autoplot(resampling, task, 1)
autoplot(resampling, task, fold_id = 2, repeats_id = 2)
autoplot(resampling, task, c(1, 2, 3, 4))
}
##########
# RepeatedSpCVEnv
##########
\dontrun{
library(mlr3)
task = tsk("ecuador")
resampling = rsmp("repeated-spcv-env", folds = 10, repeats = 2)
resampling$instantiate(task)

autoplot(resampling, task)
autoplot(resampling, task, 1)
autoplot(resampling, task, fold_id = 2, repeats_id = 2)
autoplot(resampling, task, c(1, 2, 3, 4))
}
##########
# RepeatedSpCVBlock
##########
\dontrun{
library(mlr3)
task = tsk("ecuador")
resampling = rsmp("repeated-spcv-block",
  folds = 5, repeats = 2,
  range = c(1000, 1500)
)
resampling$instantiate(task)

autoplot(resampling, task)
autoplot(resampling, task, 1)
autoplot(resampling, task, fold_id = 2, repeats_id = 2)
autoplot(resampling, task, c(1, 2, 3, 4))
}
##########
# RepeatedSptCVCluto
##########
\dontrun{
library(mlr3)
library(mlr3spatiotempcv)
task_st = tsk("cookfarm")
resampling = rsmp("repeated-sptcv-cluto", folds = 5, repeats = 2)
resampling$instantiate(task_st, "Date")

# plot
autoplot(resampling, task_st)
autoplot(resampling, task_st, 1, repeats_id = 2)
autoplot(resampling, task_st, c(1, 2, 3, 4))
}
##########
# SpCVBuffer
##########
\dontrun{
library(mlr3)
resampling = rsmp("spcv-buffer", theRange = 1000)
resampling$instantiate(task)

autoplot(resampling, task, 1)
autoplot(resampling, task, c(1, 2, 3, 4))
}
##########
# SptCVCstf
##########
\dontrun{
library(mlr3)
library(mlr3spatiotempcv)
task_st = tsk("cookfarm")
resampling = rsmp("sptcv-cstf", folds = 5)
resampling$instantiate(task_st, "Date")

# plot
autoplot(resampling, task_st)
autoplot(resampling, task_st, 1)
autoplot(resampling, task_st, c(1, 2, 3, 4))
}
}
