% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot.R
\name{autoplot.ResamplingSpCVBlock}
\alias{autoplot.ResamplingSpCVBlock}
\alias{autoplot.ResamplingSpCVEnv}
\alias{autoplot.ResamplingSpCVBuffer}
\alias{autoplot.ResamplingSpCVCoords}
\alias{autoplot.ResamplingCV}
\alias{autoplot.ResamplingRepeatedCV}
\alias{autoplot.ResamplingRepeatedSpCVCoords}
\alias{autoplot.ResamplingRepeatedSpCVEnv}
\alias{autoplot.ResamplingRepeatedSpCVBlock}
\title{Visualization of spatial resampling.}
\usage{
\method{autoplot}{ResamplingSpCVBlock}(
  object,
  task,
  fold_id = NULL,
  grid = TRUE,
  train_color = "#0072B5",
  test_color = "#E18727",
  ...
)

\method{autoplot}{ResamplingSpCVEnv}(
  object,
  task,
  fold_id = NULL,
  grid = TRUE,
  train_color = "#0072B5",
  test_color = "#E18727",
  ...
)

\method{autoplot}{ResamplingSpCVBuffer}(
  object,
  task,
  fold_id = NULL,
  grid = TRUE,
  train_color = "#0072B5",
  test_color = "#E18727",
  ...
)

\method{autoplot}{ResamplingSpCVCoords}(
  object,
  task,
  fold_id = NULL,
  grid = TRUE,
  train_color = "#0072B5",
  test_color = "#E18727",
  ...
)

\method{autoplot}{ResamplingCV}(
  object,
  task,
  fold_id = NULL,
  grid = TRUE,
  train_color = "#0072B5",
  test_color = "#E18727",
  ...
)

\method{autoplot}{ResamplingRepeatedCV}(
  object,
  task,
  fold_id = NULL,
  repeats_id = 1,
  grid = TRUE,
  train_color = "#0072B5",
  test_color = "#E18727",
  ...
)

\method{autoplot}{ResamplingRepeatedSpCVCoords}(
  object,
  task,
  fold_id = NULL,
  repeats_id = 1,
  grid = TRUE,
  train_color = "#0072B5",
  test_color = "#E18727",
  ...
)

\method{autoplot}{ResamplingRepeatedSpCVEnv}(
  object,
  task,
  fold_id = NULL,
  repeats_id = 1,
  grid = TRUE,
  train_color = "#0072B5",
  test_color = "#E18727",
  ...
)

\method{autoplot}{ResamplingRepeatedSpCVBlock}(
  object,
  task,
  fold_id = NULL,
  repeats_id = 1,
  grid = TRUE,
  train_color = "#0072B5",
  test_color = "#E18727",
  ...
)
}
\arguments{
\item{object}{\link{Resampling}\cr
mlr3 spatial resampling object. One of class \link{ResamplingSpCVBuffer},
\link{ResamplingSpCVBlock}, \link{ResamplingSpCVCoords}, \link{ResamplingSpCVEnv}.}

\item{task}{\link{TaskClassifST}/\link{TaskRegrST}\cr
mlr3 task object.}

\item{fold_id}{\link{numeric}\cr
Fold IDs to plot.}

\item{grid}{(\code{logical(1)})\cr Should a gridded plot using
\code{cowplot::plot_grid()} be created? If \code{FALSE} only a list with all
\CRANpkg{ggplot2} resamplings is returned. Default is \code{TRUE}. Only applies
if a numeric vector is passed to argument \code{fold_id}.}

\item{train_color}{\code{character(1)}\cr
The color to use for the training set observations.}

\item{test_color}{\code{character(1)}\cr
The color to use for the test set observations.}

\item{...}{Not used.}

\item{repeats_id}{\link{numeric}\cr
Repetition ID to plot.}
}
\value{
\code{\link[=ggplot]{ggplot()}} or list of ggplot2 objects.
}
\description{
Generates plots for \link{ResamplingSpCVBlock}, \link{ResamplingSpCVEnv},
\link{ResamplingSpCVBuffer} and \link{ResamplingSpCVCoords}.
}
\details{
By default a plot is returned; if \code{fold_id} is set, a gridded plot
is created. If \code{grid = FALSE}, only a list of ggplot2 resamplings is
returned. This gives the option to align the single plots manually.

When no single fold is selected, the \code{\link[ggsci:scale_color_ucscgb]{ggsci::scale_color_ucscgb()}} palette
is used to display all partitions. If you want to change the colors,
call \verb{<plot> + <color-palette>()}
}
\examples{
#####
# SpCVBlock
#####
library(mlr3)
task = tsk("ecuador")
resampling = rsmp("spcv-block", range = 1000)
resampling$instantiate(task)

## Visualize all partitions
autoplot(resampling, task)

## Visualize the train/test split of a single fold
autoplot(resampling, task, fold_id = 1)

## Visualize train/test splits of multiple folds
autoplot(resampling, task, fold_id = c(1, 2, 3, 4))

# return only a list of ggplot2 resamplings
plot_list = autoplot(resampling, task,
  fold_id = c(1, 2, 3, 4), grid = FALSE)
#####
# SpCVEnv
#####
library(mlr3)
resampling = rsmp("spcv-env", folds = 4, features = c("dem"))
resampling$instantiate(task)
autoplot(resampling, task)
autoplot(resampling, task, 1)
autoplot(resampling, task, c(1, 2, 3, 4))
#####
# SpCVBuffer
#####
library(mlr3)
resampling = rsmp("spcv-buffer", range = 1000)
resampling$instantiate(task)
autoplot(resampling, task, 1)
autoplot(resampling, task, c(1, 2, 3, 4))
#####
# SpCVCoords
#####
library(mlr3)
resampling = rsmp("spcv-coords")
resampling$instantiate(task)
autoplot(resampling, task)
autoplot(resampling, task, 1)
autoplot(resampling, task, c(1, 2, 3, 4))
#####
# Non-Spatial CV
#####
library(mlr3)
resampling = rsmp("cv")
resampling$instantiate(task)
autoplot(resampling, task)
autoplot(resampling, task, 1)
autoplot(resampling, task, c(1, 2, 3, 4))
#####
# Repeated Non-Spatial CV
#####
library(mlr3)
resampling = rsmp("repeated_cv", folds = 5, repeats = 2)
resampling$instantiate(task)
autoplot(resampling, task)
autoplot(resampling, task, 1)
autoplot(resampling, task, fold_id = 2, repeats_id = 2)
autoplot(resampling, task, c(1, 2, 3, 4))
#####
# RepeatedSpCVCoords
#####
library(mlr3)
task = tsk("diplodia")
resampling = rsmp("repeated-spcv-coords", folds = 10, repeats = 2)
resampling$instantiate(task)
autoplot(resampling, task)
autoplot(resampling, task, 1)
autoplot(resampling, task, fold_id = 2, repeats_id = 2)
autoplot(resampling, task, c(1, 2, 3, 4))
#####
# RepeatedSpCVEnv
#####
library(mlr3)
task = tsk("ecuador")
resampling = rsmp("repeated-spcv-env", folds = 10, repeats = 2)
resampling$instantiate(task)
autoplot(resampling, task)
autoplot(resampling, task, 1)
autoplot(resampling, task, fold_id = 2, repeats_id = 2)
autoplot(resampling, task, c(1, 2, 3, 4))
#####
# RepeatedSpCVBlock
#####
library(mlr3)
task = tsk("ecuador")
resampling = rsmp("repeated-spcv-block", folds = 5, repeats = 2,
  range = c(1000, 1500))
resampling$instantiate(task)
autoplot(resampling, task)
autoplot(resampling, task, 1)
autoplot(resampling, task, fold_id = 2, repeats_id = 2)
autoplot(resampling, task, c(1, 2, 3, 4))
}
