% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResamplingSpCVEnv.R
\name{ResamplingSpCVEnv}
\alias{ResamplingSpCVEnv}
\title{Environmental Block Cross Validation Resampling}
\format{\link[R6:R6Class]{R6::R6Class} inheriting from \link{Resampling}.}
\description{
Environmental Block Cross Validation. This strategy uses k-means
clustering to specify blocks of smilar environmental conditions. Only numeric
features can be used. The \code{features} used for building blocks can be
specified in the \code{param_set}. By default, all numeric features are used.
}
\section{Fields}{
 See \link{Resampling}.
}

\section{Methods}{
 See \link{Resampling}.
}

\examples{
library(mlr3)
task = tsk("ecuador")

# Instantiate Resampling
rcv = rsmp("spcv-env")
rcv$param_set$values = list(folds = 4)
rcv$instantiate(task)

# Individual sets:
rcv$train_set(1)
rcv$test_set(1)
intersect(rcv$train_set(1), rcv$test_set(1))

# Internal storage:
rcv$instance
}
\references{
Valavi R, Elith J, Lahoz-Monfort JJ, Guillera-Arroita G. blockCV:
An r package for generating spatially or environmentally separated folds for
k-fold cross-validation of species distribution models. Methods Ecol Evol.
2019; 10:225–232. https://doi.org/10.1111/2041-210X.13107
}
\section{Super class}{
\code{\link[mlr3:Resampling]{mlr3::Resampling}} -> \code{ResamplingSpCVEnv}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{ResamplingSpCVEnv$new()}}
\item \href{#method-instantiate}{\code{ResamplingSpCVEnv$instantiate()}}
\item \href{#method-clone}{\code{ResamplingSpCVEnv$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Resampling" data-id="format">}\href{../../mlr3/html/Resampling.html#method-format}{\code{mlr3::Resampling$format()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Resampling" data-id="print">}\href{../../mlr3/html/Resampling.html#method-print}{\code{mlr3::Resampling$print()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Resampling" data-id="test_set">}\href{../../mlr3/html/Resampling.html#method-test_set}{\code{mlr3::Resampling$test_set()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="mlr3" data-topic="Resampling" data-id="train_set">}\href{../../mlr3/html/Resampling.html#method-train_set}{\code{mlr3::Resampling$train_set()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResamplingSpCVEnv$new(id = "spcv-env")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-instantiate"></a>}}
\subsection{Method \code{instantiate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResamplingSpCVEnv$instantiate(task)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{ResamplingSpCVEnv$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
