% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ResamplingSpCVEnv.R
\docType{data}
\name{ResamplingSpCVEnv}
\alias{ResamplingSpCVEnv}
\title{Environmental Block Cross Validation Resampling}
\format{\link[R6:R6Class]{R6::R6Class} inheriting from \link{Resampling}.}
\usage{
ResamplingSpCVEnv
}
\description{
Environmental Block Cross Validation. This strategy uses k-means clustering to specify blocks of smilar environmental conditions.
Only numeric features can be used. The \code{features} used for building blocks can be specified in the \code{param_set}. By default, all numeric features are used.
}
\section{Fields}{

See \link{Resampling}.
}

\section{Methods}{

See \link{Resampling}.
}

\examples{
task <- mlr3::mlr_tasks$get("ecuador")

# Instantiate Resampling
rcv <- mlr3::mlr_resamplings$get("spcv-env")
rcv$param_set$values <- list(folds = 4))
rcv$instantiate(task)

# Individual sets:
rcv$train_set(1)
rcv$test_set(1)
intersect(rcv$train_set(1), rcv$test_set(1))

# Internal storage:
rcv$instance
}
\references{
Valavi R, Elith J, Lahoz-Monfort JJ, Guillera-Arroita G. blockCV: An r package for generating spatially or environmentally separated folds for k-fold cross-validation of species distribution models. Methods Ecol Evol. 2019; 10:225–232. https://doi.org/10.1111/2041-210X.13107
}
\keyword{datasets}
