% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaskClassifST.R
\docType{data}
\name{TaskClassifST}
\alias{TaskClassifST}
\title{SpatioTemporal Classification Task}
\format{\link[R6:R6Class]{R6::R6Class} object inheriting from
\link{Task}/\link{TaskSupervised}/\link{TaskClassif}.}
\description{
This task specializes \link{Task} and \link{TaskSupervised} for
spatio-temporal classification problems. The target column is assumed to be a
factor. The \code{task_type} is set to \code{"classif"} and \code{"spatiotemporal"}.

A spatial example task is available via \code{mlr_tasks$get("ecuador")}. During
initialization, coordinates need to be passed. By default, coordinates are
not used as features. This can be changed by setting \code{coords_as_features = TRUE}.
}
\section{Construction}{
\preformatted{t = TaskClassifST$new(id, backend, target, coordinates, positive = NULL,
  coords_as_features = FALSE)
}
\itemize{
\item \code{id} :: \code{character(1)}\cr
Identifier for the task.
\item \code{backend} :: \link{DataBackend}\cr Either a \link{DataBackend}, or any object which
is convertible to a DataBackend with \code{as_data_backend()}. E.g., a
\code{data.frame()} will be converted to a \link{DataBackendDataTable}.
\item \code{target} :: \code{character(1)}\cr
Name of the target column.
\item \code{coordinates} :: \link{data.frame}\cr
A \link{data.frame} with \code{ncol(2)} containing longitude and latitude information.
\item \code{positive} :: \code{character(1)}\cr
Only for binary classification: Name of the positive class.
\item \code{coords_as_features} :: \code{logical(1)}\cr
Whether the coordinates should also be used as features. Default is \code{FALSE}.
\item \code{crs} :: \code{character(1)}\cr
Coordinates reference system
}
}

\section{Fields}{

All methods from \link{TaskSupervised} and \link{TaskClassif}, and additionally:
\itemize{
\item \code{coordinates} :: \link{data.frame}\cr
Stores the coordinates passed during construction.
}
}

\section{Methods}{

See \link{TaskSupervised} and \link{TaskClassif}.
}

\examples{
data = data.table::as.data.table(readRDS(system.file("extdata", "ecuador.rda",
  package = "mlr3spatiotemporal")))
task = TaskClassifST$new("ecuador", backend = data, target = "slides",
  positive = "TRUE", coordinates = c("x", "y"))

task$task_type
task$formula()
task$truth()
task$class_names
task$positive
task$negative
task$coordinates()
task$coordinate_names

# possible properties:
mlr3::mlr_reflections$task_properties$classif
}
\seealso{
Example spatial classification task: \code{\link[=mlr_tasks_ecuador]{ecuador}}.

Other Task: \code{\link{TaskRegrST}}
}
\concept{Task}
\keyword{datasets}
