% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaskRegrST.R
\docType{data}
\name{TaskRegrST}
\alias{TaskRegrST}
\title{SpatioTemporal Regression Task}
\format{\link[R6:R6Class]{R6::R6Class} object inheriting from \link{Task}/\link{TaskSupervised}/\link{TaskRegr}.}
\description{
This task specializes \link{Task} and \link{TaskSupervised} for spatio-temporal regression problems.
The target column is assumed to be a factor.
The \code{task_type} is set to \code{"classif"} and \code{"spatiotemporal"}.

A spatial example task is available via \code{mlr_tasks$get("ecuador")}.
During initialization, coordinates need to be passed.
By default, coordinates are not used as features.
This can be changed by setting \code{coords_as_features = TRUE}.
}
\section{Construction}{
\preformatted{t = TaskRegrST$new(id, backend, target, coordinates,
  coords_as_features = FALSE)
}
\itemize{
\item \code{id} :: \code{character(1)}\cr
Identifier for the task.
\item \code{backend} :: \link{DataBackend}\cr
Either a \link{DataBackend}, or any object which is convertible to a DataBackend with \code{as_data_backend()}.
E.g., a \code{data.frame()} will be converted to a \link{DataBackendDataTable}.
\item \code{target} :: \code{character(1)}\cr
Name of the target column.
\item \code{coordinates} :: \link{data.frame}\cr
A \link{data.frame} with \code{ncol(2)} containing longitude and latitude information.
\item \code{coords_as_features} :: \code{logical(1)}\cr
Whether the coordinates should also be used as features. Default is \code{FALSE}.
\item \code{crs} :: \code{character(1)}\cr
Coordinates reference system
}
}

\section{Fields}{

All methods from \link{TaskSupervised} and \link{TaskRegr}, and additionally:
\itemize{
\item \code{coordinates} :: \link{data.frame}\cr
Stores the coordinates passed during construction.
}
}

\section{Methods}{

See \link{TaskSupervised} and \link{TaskRegr}.
}

\examples{

# possible properties:
mlr3::mlr_reflections$task_properties$regr
}
\seealso{
Example spatial regression task: \code{\link[=mlr_tasks_ecuador]{ecuador}}.

Other Task: \code{\link{TaskClassifST}}
}
\concept{Task}
\keyword{datasets}
